Add stage3 for consolidated outputs across cohorts
